Add a GameState class
This is an abstract class representing the current view of the game.
The game is designed to have several states, at least:
  * an intro state,
  * a main menu state,
  * an in-game state
and possibly others. Each state is a class providing three mandatory
functions: a way to handle user's input, a method to apply
state-specific game logic and a subroutine to render the scene.