Fix bug in FindChains implementation, extend tests
Fixes a bug in FindChains() method - one that surprisingly has not been
detected via unit tests, and which resulted with chains being marked,
but function that found them still returning "no chains found" result.
Extends the tests to include a board generated with flawed version of
the program. The fixed implementation passes all the tests.